# server/docs/swagger.yaml
basePath: /
definitions:
  users.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  users.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  users.Role:
    enum:
    - therapist
    - assistant
    - patient
    type: string
    x-enum-varnames:
    - RoleTherapist
    - RoleAssistant
    - RolePatient
  users.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastName:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      phoneNumber:
        type: string
      role:
        $ref: '#/definitions/users.Role'
      therapistID:
        description: null for therapists
        type: string
      updatedAt:
        type: string
    type: object
  users.UserRequest:
    properties:
      email:
        type: string
      lastName:
        type: string
      name:
        type: string
      password:
        example: secret123
        minLength: 6
        type: string
      phoneNumber:
        type: string
      role:
        $ref: '#/definitions/users.Role'
    required:
    - email
    - lastName
    - name
    - password
    - phoneNumber
    - role
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API server for therapist-patient management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Chrono API
  version: "1.0"
paths:
  /users:
    get:
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.User'
            type: array
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
    post:
      description: Create a new user
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.User'
      summary: Create a user
      tags:
      - users
  /users/{id}:
    get:
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - users
  /users/login:
    post:
      description: Authenticate user and return JWT token
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/users.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /users/refresh-token:
    post:
      description: Refresh JWT access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/users.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /users/register:
    post:
      description: Register a new user account
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by your JWT token, e.g. "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"